Added features to bljoin and fixed a bug.
Added default behavior: throw an error if two records in the same file
have the same ID. Added an option to turn that behavior off.

Added an option to separate each original record by a user-defined
string (e.g. a bunch of Ns).

Fixed bug with padding sequences not found in every file.